requirements.txt
semantic-kernel==1.25.0
azure-core>=1.31.0
azure-identity>=1.18.0
azure-cosmos>=3.2.0
azure-monitor-opentelemetry>=1.6.5
botbuilder-azure>=4.16.1
botbuilder-core>=4.17.0
botbuilder-dialogs>=4.16.1
botbuilder-integration-aiohttp>=4.16.1
python-dotenv>=1.0.1
openai>=1.61.0
black>=23.3.0
ruff>=0.1.7
mypy>=1.0.0
pydoclint>=0.6.0
pymarkdownlnt>=0.9.0
yamllint>=1.32.0
isort>=6.0.1
